+++ src/FieldDefinition.php
@@ -20,6 +20,11 @@ final class FieldDefinition
         return FieldDefinition\Closure::required($closure);
     }
 
+    public static function optionalClosure(\Closure $closure): FieldDefinition\Closure
+    {
+        return FieldDefinition\Closure::optional($closure);
+    }
+
     /**
      * @phpstan-param class-string<T> $className
      * @phpstan-return FieldDefinition\Reference<T>
@@ -36,6 +41,24 @@ final class FieldDefinition
     public static function reference(string $className): FieldDefinition\Reference
     {
         return FieldDefinition\Reference::required($className);
+    }
+
+    /**
+     * @phpstan-param class-string<T> $className
+     * @phpstan-return FieldDefinition\Reference<T>
+     * @phpstan-template T
+     *
+     * @psalm-param class-string<T> $className
+     * @psalm-return FieldDefinition\Reference<T>
+     * @psalm-template T
+     *
+     * @param string $className
+     *
+     * @return FieldDefinition\Reference
+     */
+    public static function optionalReference(string $className): FieldDefinition\Reference
+    {
+        return FieldDefinition\Reference::optional($className);
     }
 
     /**
@@ -63,6 +86,30 @@ final class FieldDefinition
     }
 
     /**
+     * @phpstan-param class-string<T> $className
+     * @phpstan-return FieldDefinition\References<T>
+     * @phpstan-template T
+     *
+     * @psalm-param class-string<T> $className
+     * @psalm-return FieldDefinition\References<T>
+     * @psalm-template T
+     *
+     * @param string $className
+     * @param int    $count
+     *
+     * @throws Exception\InvalidCount
+     *
+     * @return FieldDefinition\References
+     */
+    public static function optionalReferences(string $className, int $count = 1): FieldDefinition\References
+    {
+        return FieldDefinition\References::optional(
+            $className,
+            $count
+        );
+    }
+
+    /**
      * @param string $value
      * @param int    $initialNumber
      *
@@ -75,6 +122,24 @@ final class FieldDefinition
         }
 
         return FieldDefinition\Sequence::required(
+            $value,
+            $initialNumber
+        );
+    }
+
+    /**
+     * @param string $value
+     * @param int    $initialNumber
+     *
+     * @return FieldDefinition\Sequence
+     */
+    public static function optionalSequence($value, int $initialNumber = 1): FieldDefinition\Sequence
+    {
+        if (false === \strpos($value, '%d')) {
+            $value .= '%d';
+        }
+
+        return FieldDefinition\Sequence::optional(
             $value,
             $initialNumber
         );
@@ -96,5 +161,23 @@ final class FieldDefinition
     public static function value($value): FieldDefinition\Value
     {
         return FieldDefinition\Value::required($value);
+    }
+
+    /**
+     * @phpstan-param T $value
+     * @phpstan-return FieldDefinition\Value<T>
+     * @phpstan-template T
+     *
+     * @psalm-param T $value
+     * @psalm-return FieldDefinition\Value<T>
+     * @psalm-template T
+     *
+     * @param mixed $value
+     *
+     * @return FieldDefinition\Value
+     */
+    public static function optionalValue($value): FieldDefinition\Value
+    {
+        return FieldDefinition\Value::optional($value);
     }
 }