+++ test/Unit/FieldDefinition/ClosureTest.php
@@ -30,7 +30,7 @@ use Ergebnis\FactoryBot\Test\Unit\AbstractTestCase;
  */
 final class ClosureTest extends AbstractTestCase
 {
-    public function testResolveReturnsResultOfInvokingClosureWithFixtureFactory(): void
+    public function testRequiredResolvesToResultOfInvokingClosureWithFixtureFactory(): void
     {
         $fixtureFactory = new FixtureFactory(self::entityManager());
 
@@ -41,6 +41,27 @@ final class ClosureTest extends AbstractTestCase
         };
 
         $fieldDefinition = Closure::required($closure);
+
+        self::assertTrue($fieldDefinition->isRequired());
+
+        $resolved = $fieldDefinition->resolve($fixtureFactory);
+
+        self::assertInstanceOf(Fixture\FixtureFactory\Entity\User::class, $resolved);
+    }
+
+    public function testOptionalResolvedToResultOfInvokingClosureWithFixtureFactory(): void
+    {
+        $fixtureFactory = new FixtureFactory(self::entityManager());
+
+        $fixtureFactory->defineEntity(Fixture\FixtureFactory\Entity\User::class);
+
+        $closure = static function (FixtureFactory $fixtureFactory): Fixture\FixtureFactory\Entity\User {
+            return $fixtureFactory->create(Fixture\FixtureFactory\Entity\User::class);
+        };
+
+        $fieldDefinition = Closure::optional($closure);
+
+        self::assertFalse($fieldDefinition->isRequired());
 
         $resolved = $fieldDefinition->resolve($fixtureFactory);