+++ test/Unit/FieldDefinition/ReferencesTest.php
@@ -37,7 +37,53 @@ final class ReferencesTest extends AbstractTestCase
      *
      * @param int $count
      */
-    public function testConstructorRejectsInvalidCount(int $count): void
+    public function testOptionalRejectsInvalidCount(int $count): void
+    {
+        $this->expectException(Exception\InvalidCount::class);
+        $this->expectExceptionMessage(\sprintf(
+            'Count needs to be greater than or equal to 1, but %d is not.',
+            $count
+        ));
+
+        References::optional(
+            Fixture\FixtureFactory\Entity\User::class,
+            $count
+        );
+    }
+
+    /**
+     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\NumberProvider::intGreaterThanOne()
+     *
+     * @param int $count
+     */
+    public function testOptionalResolvesToArrayOfObjectsCreatedByFixtureFactory(int $count): void
+    {
+        $className = Fixture\FixtureFactory\Entity\User::class;
+
+        $fixtureFactory = new FixtureFactory(self::entityManager());
+
+        $fixtureFactory->defineEntity($className);
+
+        $fieldDefinition = References::optional(
+            $className,
+            $count
+        );
+
+        self::assertFalse($fieldDefinition->isRequired());
+
+        $resolved = $fieldDefinition->resolve($fixtureFactory);
+
+        self::assertIsArray($resolved);
+        self::assertCount($count, $resolved);
+        self::assertContainsOnly($className, $resolved);
+    }
+
+    /**
+     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\NumberProvider::intLessThanOne()
+     *
+     * @param int $count
+     */
+    public function testRequiredRejectsInvalidCount(int $count): void
     {
         $this->expectException(Exception\InvalidCount::class);
         $this->expectExceptionMessage(\sprintf(
@@ -56,7 +102,7 @@ final class ReferencesTest extends AbstractTestCase
      *
      * @param int $count
      */
-    public function testResolveReturnsObjectsFromFixtureFactory(int $count): void
+    public function testRequiredResolvesToArrayOfObjectsCreatedByFixtureFactory(int $count): void
     {
         $className = Fixture\FixtureFactory\Entity\User::class;
 
@@ -69,8 +115,11 @@ final class ReferencesTest extends AbstractTestCase
             $count
         );
 
+        self::assertTrue($fieldDefinition->isRequired());
+
         $resolved = $fieldDefinition->resolve($fixtureFactory);
 
+        self::assertIsArray($resolved);
         self::assertCount($count, $resolved);
         self::assertContainsOnly($className, $resolved);
     }